fix: always add trailing Slash
Using `trailingSlash: 'ignore'` will not generates redirect with page
end or not by /

For Cloudflare, if no page found, /foo == /foo/ == /foo/index.html. And
no redirect occurs foo/index.html will be serve for all of them.

We need to be explicit and choice which type of url we prefer and gatsby
will generates the appropriate pages/code to redirect all other to our
prefered one.

So we need to choice trailingSlash always or never. I pick `always` like
we did for the old documentation.